py/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==3.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
Bottleneck==1.3.5
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.4
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
environs==9.5.0
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
Flask==2.2.5
Flask-Cors==4.0.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.12.2
google-auth==2.26.2
googleapis-common-protos==1.62.0
greenlet==3.0.1
grpcio==1.58.0
h11==0.14.0
h2==4.0.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.0.0
InstructorEmbedding==1.0.1
itsdangerous==2.0.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.0
kubernetes==29.0.0
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.21
langsmith==0.0.87
MarkupSafe==2.1.3
marshmallow==3.20.2
minio==7.2.3
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nlp==0.4.0
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.7.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.6.0
packaging==23.2
pandas==2.1.4
pandas-stubs==2.1.4.231227
pillow==10.2.0
pip==23.3.1
pluggy==1.4.0
posthog==3.3.2
protobuf==4.25.2
pulsar-client==3.4.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.12
pymilvus==2.3.5
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pytest==8.0.0
pytest-subtests==0.11.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.4.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==2.2.0
tokenizers==0.15.1
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.65.0
transformers==4.37.1
typer==0.9.0
types-pytz==2022.4.0.0
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.3
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.27.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.2.3
wheel==0.41.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
